Allows customization of destination dotfile
The most common places from which to source the sd function definition
are either in .bashrc or .bash_profile, but which one is desired depends
on your environment and personal preferences. Therefore, I'm going to
stick to the default of .bash_profile, but allow one to pass an
additional argument to either 'upgrade' or 'install' that defines where
the 'source sd' call is going.